third_party/WebKit/LayoutTests/inspector/console/console-log-syntax-error.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/console-test.js"></script>
<script src="resources/syntax-error.js"></script>
<script>
function test()
{
    InspectorTest.dumpConsoleMessages(undefined, undefined, simpleFormatter);
    InspectorTest.completeTest();
    function simpleFormatter(element, message)
    {
    	return message.messageText + ":" + message.line + ":" + message.column;
    }
}
</script>
</head>
<body onload="runTest()">
<p>
Tests that syntax errors are logged into console and doesn't cause browser crash.
</p>
</body>
</html>
