third_party/WebKit/LayoutTests/inspector/console/console-log-syntax-error-expected.txt

CONSOLE ERROR: line 3: Uncaught SyntaxError: Unexpected token )
Tests that syntax errors are logged into console and doesn't cause browser crash.

Uncaught SyntaxError: Unexpected token ):3:17
